+++ e2e/pkg/fixtures/ingress.go
@@ -4,7 +4,7 @@ import (
 	"os"
 	"strconv"
 
-	"github.com/h3poteto/aws-global-accelerator-controller/pkg/apis"
+	apis "github.com/h3poteto/aws-global-accelerator-controller/pkg/apis/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 	networkingv1 "k8s.io/api/networking/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
