+++ e2e/pkg/fixtures/service.go
@@ -1,7 +1,7 @@
 package fixtures
 
 import (
-	"github.com/h3poteto/aws-global-accelerator-controller/pkg/apis"
+	apis "github.com/h3poteto/aws-global-accelerator-controller/pkg/apis/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
